pyproject.toml
[project]
name = "divisions"
version = "0.1.0"
description = "A curated, hand-verified dataset of OCD (Open Civic Data) divisions in YAML."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "TBD" }
authors = [
  { name = "VoterBloc" },
]
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "click>=8.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[project.scripts]
divisions = "divisions.cli:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/divisions"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
